refactor(mckmaps): merges renderContext into template
in addition, changes the structure of sourcecode files
