+++ internal/mckmaps/filter.go
@@ -3,7 +3,6 @@ package mckmaps
 import (
 	"errors"
 	"io/ioutil"
-	"path/filepath"
 	"regexp"
 	"strings"
 
@@ -18,6 +17,24 @@ var (
 	ppParseRegexp    = makeParseRegexpFilter()
 	ppToJSONMatcher  = &jsonMatcherFilter{}
 )
+
+func makeTemplateFilter() *templateFilter {
+	f := templateFilter{}
+	f.reset()
+	return &f
+}
+
+func makeLoadFileFilter() *loadFileFilter {
+	f := new(loadFileFilter)
+	f.reset()
+	return f
+}
+
+func makeParseRegexpFilter() *parseRegexpFilter {
+	f := new(parseRegexpFilter)
+	f.reset()
+	return f
+}
 
 func doFiltersOnV(v interface{}, filters ...filter) (interface{}, error) {
 	chain := &filterChain{filters: filters}
@@ -72,173 +89,8 @@ func (f *commentFilter) removeComment(v interface{}) {
 	}
 }
 
-type templateFilter struct { // rewrites @template directives with given vars
-	templateCache  map[string]*template
-	varsSliceCache map[string][]*vars
-}
-
-func makeTemplateFilter() *templateFilter {
-	f := templateFilter{}
-	f.reset()
-	return &f
-}
-
-func (f *templateFilter) doFilter(v interface{}, chain *filterChain) error {
-	rV, err := f.rewrite(v)
-	if err != nil {
-		return err
-	}
-	return chain.doFilter(rV)
-}
-
-func (f *templateFilter) rewrite(v interface{}) (interface{}, error) {
-	var rV interface{}
-	var err error
-	switch v.(type) {
-	case myjson.Object:
-		rV, err = f.rewriteObject(v.(myjson.Object))
-	case myjson.Array:
-		rV, err = f.rewriteArray(v.(myjson.Array))
-	default:
-		rV, err = v, nil
-	}
-	return rV, err
-}
-
-func (f *templateFilter) rewriteObject(v myjson.Object) (interface{}, error) {
-	if v.Has(dTemplate) { // if v is a @template directive
-		template, ctx, err := f.getTemplateFromDTemplate(v)
-		if err != nil {
-			return nil, err
-		}
-		varsSlice, err := f.getVarsFromDTemplate(v)
-		if err != nil {
-			return nil, err
-		}
-
-		rV, err := template.render(ctx, varsSlice)
-		if err != nil {
-			return nil, err
-		}
-		return fromTemplate{rV: rV}, nil
-	} else {
-		result := make(myjson.Object, len(v))
-		for name, value := range v {
-			rValue, err := f.rewrite(value)
-			if err != nil {
-				return nil, err
-			}
-
-			switch rValue.(type) {
-			case fromTemplate:
-				rValue = rValue.(fromTemplate).forObject()
-			}
-
-			result[name] = rValue
-		}
-		return result, nil
-	}
-}
-
-func (f *templateFilter) rewriteArray(v myjson.Array) (myjson.Array, error) {
-	var result myjson.Array
-	for _, value := range v {
-		rValue, err := f.rewrite(value)
-		if err != nil {
-			return nil, err
-		}
-
-		switch rValue.(type) {
-		case fromTemplate:
-			for _, _rValue := range rValue.(fromTemplate).forArray() {
-				result = append(result, _rValue)
-			}
-		default:
-			result = append(result, rValue)
-		}
-	}
-	return result, nil
-}
-
-func (f *templateFilter) getTemplateFromDTemplate(v myjson.Object) (*template, *renderContext, error) {
-	filename, err := v.GetString(dTemplate)
-	if err != nil {
-		return nil, nil, errors.New("cannot read the name of template file")
-	}
-
-	var template *template
-	var ok bool
-	_filename := string(filename)
-	if template, ok = f.templateCache[_filename]; !ok {
-		tParser := &templateParser{Parser: Parser{filename: _filename}}
-		template, err = tParser.parse()
-		if err != nil {
-			return nil, nil, err
-		}
-		f.templateCache[_filename] = template
-	}
-	return template, &renderContext{filename: _filename}, nil
-}
-
-func (f *templateFilter) getVarsFromDTemplate(v myjson.Object) (varsSlice []*vars, err error) {
-	if v.Has(tVars) { // if @template directive has a 'vars' attribute
-		varsSlice, err = new(varsJSONParser).parseVars(v)
-	} else {
-		var filename myjson.String
-		filename, err = v.GetString(dVars)
-		if err != nil {
-			err = errors.New("cannot read filename from " + dVars)
-			return
-		}
-
-		var ok bool
-		_filename := string(filename)
-		if varsSlice, ok = f.varsSliceCache[_filename]; !ok {
-			ext := filepath.Ext(_filename)
-			if ext == ".csv" {
-				vParser := &varsCSVParser{Parser: Parser{filename: _filename}}
-				varsSlice, err = vParser.parse()
-			} else {
-				vParser := &varsJSONParser{Parser: Parser{filename: _filename}}
-				varsSlice, err = vParser.parse()
-			}
-			f.varsSliceCache[_filename] = varsSlice
-		}
-	}
-	return
-}
-
-func (f *templateFilter) reset() { // clears caches
-	f.templateCache = make(map[string]*template)
-	f.varsSliceCache = make(map[string][]*vars)
-}
-
-type fromTemplate struct {
-	rV myjson.Array
-}
-
-func (ft fromTemplate) forObject() interface{} {
-	if len(ft.rV) == 0 {
-		return nil
-	} else if len(ft.rV) == 1 {
-		return ft.rV[0]
-	} else {
-		return ft.rV
-	}
-}
-
-func (ft fromTemplate) forArray() []interface{} {
-	return ft.rV
-}
-
 type loadFileFilter struct { // loads file contents for @file
 	fileCache map[string][]byte
-}
-
-func makeLoadFileFilter() *loadFileFilter {
-	f := new(loadFileFilter)
-	f.reset()
-	return f
 }
 
 func (f *loadFileFilter) doFilter(v interface{}, chain *filterChain) error {
@@ -317,12 +169,6 @@ func (f *loadFileFilter) reset() {
 
 type parseRegexpFilter struct {
 	regexpCache map[string]myjson.ExtRegexp
-}
-
-func makeParseRegexpFilter() *parseRegexpFilter {
-	f := new(parseRegexpFilter)
-	f.reset()
-	return f
 }
 
 func (f *parseRegexpFilter) doFilter(v interface{}, chain *filterChain) error {
