--- internal/myjson/testdata/json1-slim.json
@@ -1,1 +1,0 @@
-{"arr":[1,"a",null],"bool":false,"null":null,"num":1.23,"str":"hello"}
